refactor(omt): move omt from plugin to native registration
Removes OmtPlugin wrapper and registers omt tools directly in the tool
registry. System prompt injection (notepad priority, project memory) now
runs as native Effect code in both the agent layer and LLM request prep,
eliminating the plugin hook indirection.
